+++ server/src/main/java/com/dialog/server/dto/auth/request/SignInRequest.java
@@ -1,4 +1,4 @@
-package com.dialog.server.dto;
+package com.dialog.server.dto.auth.request;
 
 public record SignInRequest(String oid) {
 }
