+++ server/src/main/java/com/dialog/server/dto/auth/response/TempUserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.dialog.server.dto;
+package com.dialog.server.dto.auth.response;
 
 public record TempUserInfoResponse(String email) {
 }
