+++ server/src/main/java/com/dialog/server/dto/security/GitHubOAuth2UserInfo.java
@@ -1,4 +1,4 @@
-package com.dialog.server.common.security.domain;
+package com.dialog.server.dto.security;
 
 import com.dialog.server.exception.DialogException;
 import com.dialog.server.exception.ErrorCode;
