+++ server/src/main/java/com/dialog/server/service/AuthService.java
@@ -1,7 +1,7 @@
 package com.dialog.server.service;
 
 import com.dialog.server.domain.User;
-import com.dialog.server.dto.SignupRequest;
+import com.dialog.server.dto.auth.request.SignupRequest;
 import com.dialog.server.exception.DialogException;
 import com.dialog.server.exception.ErrorCode;
 import com.dialog.server.repository.UserRepository;
